requirements.txt
customtkinter~=5.2.1
Pillow~=10.1.0